Resolve typing issues in Version
